Refactor banner
